Updated December 13, 2024
This was long over due. I have finally updated my link in bio. and I am excited at how it turned out. I did change everything I wanted it to be more simple and straight to the point. I also took out the text that was there has I left it was just not needed anymore.